remove unnecessary changes from #77903 to make changes more minimal
